+++ ISU_Medieval_Odyssey/Items/Armour/Armour.cs
@@ -50,10 +50,10 @@ namespace ISU_Medieval_Odyssey
         /// <param name="maxDefense">The maximum allowed defense on this <see cref="Armour"/></param>
         /// <param name="minDurability">The minimum allowed durability on this <see cref="Armour"/></param>
         /// <param name="maxDurability"><The minimum allowed durability on this <see cref="Armour"/>/param>
-        protected void SetArmourStatistics(int minDefense, int maxDefense, int minDurability, int maxDurability)
+        protected void SetArmourStatistics(short minDefense, short maxDefense, short minDurability, short maxDurability)
         {
             // Calculating and assigning various armour statistics
-            int durability = SharedData.RNG.Next(minDurability, maxDurability + 1);
+            short durability = (short)SharedData.RNG.Next(minDurability, maxDurability + 1);
             durabilityBar = new ProgressBar(new Rectangle(0, 0, 50, 5), durability, durability, Color.White * 0.6f, Color.Green * 0.6f);
             durabilityBar.Update();
             defence = SharedData.RNG.Next(minDefense, maxDefense + 1);
